grant

Module connection

Structs

DbConnection

Connection to the database, currently only Postgres and Redshift is supported TODO: support multiple adapters

User

Presentation for a user in the database

UserDatabaseRole

Presentation for a user database privilege in the database which a users has create or temp on database

UserSchemaRole

Presentation for a user schema privilege in the database which a users has create or usage on schema

UserTableRole

Presentation for a user table privilege in the database which a users has select, insert, update, delete or reference on table